lab2. Improve mistakes
