+++ vue/src/components/pages/GamePage.vue
@@ -50,37 +50,17 @@
           @touchend="() => onTouchEnd()"
         />
 
-        <div
-          class="car car--player"
-          :style="getCarStyle(gameState.playerCar, true)"
-        >
-          <div class="car__body">
-            <div class="car__roof" />
-            <div class="car__side car__side--left" />
-            <div class="car__side car__side--right" />
-            <div class="car__wheel car__wheel--fl" />
-            <div class="car__wheel car__wheel--fr" />
-            <div class="car__wheel car__wheel--rl" />
-            <div class="car__wheel car__wheel--rr" />
-          </div>
-        </div>
+        <GameCar
+          variant="player"
+          :style-object="getCarStyle(gameState.playerCar, true)"
+        />
 
-        <div
+        <GameCar
           v-for="enemy in gameState.enemies"
           :key="enemy.id"
-          class="car car--enemy"
-          :style="getCarStyle(enemy, false)"
-        >
-          <div class="car__body">
-            <div class="car__roof" />
-            <div class="car__side car__side--left" />
-            <div class="car__side car__side--right" />
-            <div class="car__wheel car__wheel--fl" />
-            <div class="car__wheel car__wheel--fr" />
-            <div class="car__wheel car__wheel--rl" />
-            <div class="car__wheel car__wheel--rr" />
-          </div>
-        </div>
+          variant="enemy"
+          :style-object="getCarStyle(enemy, false)"
+        />
 
         <div v-if="isPaused" class="game__overlay game__overlay--pause">
           <div class="game__overlay-text">
@@ -113,6 +93,7 @@ import { useStore } from 'vuex'
 import { useRouter } from 'vue-router'
 import { ROUTES } from '@/router'
 import { ACTIONS } from '@/store'
+import GameCar from '@/components/ui/GameCar.vue'
 
 const ROAD_WIDTH = 400
 const LANE_COUNT = 3
@@ -334,32 +315,13 @@ const spawnEnemy = () => {
 }
 
 const getCarStyle = (car: Car, isPlayer: boolean) => {
-  const scale = canvasSize.scale
-  const centerYpx = (car.y + car.height / 2) * scale
-  const bounds = getRoadBoundsAtY(centerYpx)
-  const centerXpx = bounds.leftPx + (car.x / ROAD_WIDTH) * bounds.widthPx
-
-  const gameHeight = canvasSize.height / scale
-  const depthRatio = Math.max(0, Math.min(1, (car.y + car.height / 2) / gameHeight))
-  const sizeScale = isPlayer ? 1 : 0.72 + 0.28 * depthRatio
-
-  const width = car.width * scale * sizeScale
-  const height = car.height * scale * sizeScale
-
-  const roadCenterX = bounds.leftPx + bounds.widthPx / 2
-  const roadHalfWidth = bounds.widthPx / 2
-  const offsetRatio = roadHalfWidth > 0 ? (centerXpx - roadCenterX) / roadHalfWidth : 0
-  const maxTiltDeg = 22
-  const tiltDeg = isPlayer ? offsetRatio * maxTiltDeg : 0
-
-  const baseColor = isPlayer ? '#3b82f6' : (car.color || '#9ca3af')
-
-  return {
-    width: `${width}px`,
-    height: `${height}px`,
-    transform: `translate3d(${centerXpx}px, ${centerYpx}px, 0) translate(-50%, -50%) scale(${sizeScale}) rotateX(40deg) rotateY(${tiltDeg}deg)`,
-    '--car-color': baseColor,
-  }
+  return store.getters.getCarStyle({
+    car,
+    isPlayer,
+    canvasSize,
+    roadWidth: ROAD_WIDTH,
+    tunnelTopRatio: TUNNEL_TOP_RATIO,
+  })
 }
 
 const detectCollisions = () => {
@@ -681,89 +643,6 @@ onUnmounted(() => {
   &__hint {
     width: 100%;
     max-width: 420px;
-  }
-}
-
-.car {
-  position: absolute;
-  top: 0;
-  left: 0;
-  transform-style: preserve-3d;
-  transform-origin: center center;
-  pointer-events: none;
-  transition: transform 0.12s ease-out;
-  z-index: 2;
-
-  &__body {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    border-radius: 8px;
-    background: linear-gradient(
-      180deg,
-      rgba(255, 255, 255, 0.7),
-      rgba(255, 255, 255, 0.2)
-    ),
-    var(--car-color, #9ca3af);
-    box-shadow: 0 10px 20px rgba(0, 0, 0, 0.55);
-    overflow: hidden;
-  }
-
-  &__roof {
-    position: absolute;
-    inset: 18%;
-    border-radius: 6px;
-    background: linear-gradient(180deg, #1f2937, #111827);
-    opacity: 0.9;
-  }
-
-  &__side {
-    position: absolute;
-    top: 20%;
-    bottom: 20%;
-    width: 5px;
-    background: linear-gradient(180deg, rgba(15, 23, 42, 0.9), rgba(15, 23, 42, 0.4));
-
-    &--left {
-      left: 0;
-    }
-
-    &--right {
-      right: 0;
-    }
-  }
-
-  &__wheel {
-    position: absolute;
-    width: 18%;
-    height: 14%;
-    border-radius: 999px;
-    background: #020617;
-    box-shadow: 0 2px 3px rgba(0, 0, 0, 0.6);
-
-    &--fl {
-      left: 4%;
-      top: 6%;
-    }
-
-    &--fr {
-      right: 4%;
-      top: 6%;
-    }
-
-    &--rl {
-      left: 4%;
-      bottom: 6%;
-    }
-
-    &--rr {
-      right: 4%;
-      bottom: 6%;
-    }
-  }
-
-  &--player {
-    z-index: 3;
   }
 }
